Fix .pgpass file and support additional options to entrypoint
The .pgpass file requires an * if a parameter is not defined, otherwise
it may not connect to the server.

The target database is now created if it does not already exist to
ensure it can be connected to during the copy. This requires the target
user to have sufficient permissions to create the database.

Signed-off-by: Byron Ruth <b@devel.io>